Inlined everything
